refactors files to meet go tracer standard
