+++ contrib/jackc/pgx.v5/pgxpool/example_test.go
@@ -1,10 +1,14 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2022 Datadog, Inc.
+
 package pgxpool_test
 
 import (
 	"context"
+	"gopkg.in/DataDog/dd-trace-go.v1/contrib/jackc/pgx.v5/pgxpool"
 	"log"
-
-	pgxpool "gopkg.in/DataDog/dd-trace-go.v1/contrib/jackc/pgx/v5/pgxpool"
 )
 
 func Example() {
