+++ contrib/jackc/pgx.v5/trace_test.go
@@ -1,4 +1,9 @@
-package pgx
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2022 Datadog, Inc.
+
+package pgx_v5
 
 import (
 	"context"
