+++ ddtrace/tracer/option_test.go
@@ -380,9 +380,7 @@ func TestIntegrationEnabled(t *testing.T) {
 		}
 		p := strings.Replace(pkg.Dir, pkg.Root, "../..", 1)
 		body, err := exec.Command("grep", "-rl", "MarkIntegrationImported", p).Output()
-		if err != nil {
-			t.Fatalf(err.Error())
-		}
+		require.NoErrorf(t, err, "integration: %v", pkg)
 		assert.NotEqual(t, len(body), 0, "expected %s to call MarkIntegrationImported", pkg.Name)
 	}
 }
